feat(refactoring): modules path
