+++ src/index.ts
@@ -1,4 +1,3 @@
 export * as stacks from './stacks';
-export * as constructs from './constructs';
-export * as env from './common/env';
-export * as utils from './common/utils';
+export * as aws_lambda from './constructs/aws-lambda';
+export * as common from './common';
