+++ src/stacks/index.ts
@@ -1,3 +1,3 @@
 export * from './base';
-export * as monitoring from './monitoring';
-export * as oidc from './oidc';
+export * from './newrelic';
+export * from './oidc';
