--- src/stacks/monitoring/index.ts
@@ -1,1 +1,0 @@
-export * as newrelic from './newrelic';
