--- src/stacks/oidc/index.ts
@@ -1,1 +1,0 @@
-export * as github from './github';
